backend/requirements.txt
django-filter
dj_database_url
sentry_sdk
python-json-logger
djangorestframework
djangorestframework-simplejwt
Django>=5.0
python-decouple
requests
openpyxl
numpy
whitenoise
django-cors-headers
django-redis
django-axes
django-health-check
psutil
redis
pika
django-storages
boto3
gunicorn
